Move IConsole out of Exporters and fix usings
