+++ src/Tools/dotnet-counters/Exporters/ConsoleWriter.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using Microsoft.Diagnostics.Monitoring.EventPipe;
-using Microsoft.Diagnostics.Tools.Common.Exporters;
+using Microsoft.Diagnostics.Tools.Common;
 
 namespace Microsoft.Diagnostics.Tools.Counters.Exporters
 {
