+++ src/Tools/dotnet-counters/Exporters/DefaultConsole.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
-using Microsoft.Diagnostics.Tools.Common.Exporters;
+using Microsoft.Diagnostics.Tools.Common;
 
 namespace Microsoft.Diagnostics.Tools.Counters.Exporters
 {
